ARM-2: curate for CI green — split model_a/moe under 500-LoC cap + fmt
The native-MXFP4 feature pushed two files over the ≤500-LoC cap (main→ours:
model_a.rs 492→534, moe/mod.rs 490→546; neither allow-listed). Split each into
a cohesive sibling module, re-exported so all call sites are unchanged:
  weight_map/fp8_dequant.rs   (187) — FP8-E4M3→BF16 per-tensor dequant helpers
  layers/moe/ptr_table_build.rs (135) — NVFP4/BF16/FP8 expert ptr-table builders
Leaves model_a.rs 364, moe/mod.rs 425. Also `cargo fmt` (feature files carried
unformatted `==` line-wraps from prior sessions) + drop now-unused imports.

CI mirror GREEN: cargo fmt --check, cargo clippy --workspace --tests, ≤500-LoC
gate. GB10 re-verify after the split (behavior-preserving): Leg-2 CHECK 1 256/256
maxULP 0; CHECK 2/3 bit-identical; CHECK 4 all 5 entries bit-exact; CHECK 5 clean.
