+++ crates/spark-model/src/layers/moe/helpers_a.rs
@@ -69,13 +69,12 @@ impl MoeLayer {
         // ARM-2 Phase-K Family C: native-MXFP4 routed experts have per-32 E8M0
         // scales ([N, K/32]); NVFP4 is per-16. The scale transpose must use the
         // matching block size or the E8M0 kernels read a mis-shaped scale table.
-        let routed_gs = if self.experts_scale_kind
-            == crate::weight_map::WeightQuantFormat::Mxfp4E8m0
-        {
-            32
-        } else {
-            16
-        };
+        let routed_gs =
+            if self.experts_scale_kind == crate::weight_map::WeightQuantFormat::Mxfp4E8m0 {
+                32
+            } else {
+                16
+            };
         for expert in &self.weights.experts {
             if expert.gate_proj.is_null() {
                 gate_t.push(QuantizedWeight::null());
@@ -84,10 +83,22 @@ impl MoeLayer {
                     down_t.push(QuantizedWeight::null());
                 }
             } else {
-                gate_t.push(expert.gate_proj.transpose_for_gemm_gs(gpu, inter, h, routed_gs)?);
-                up_t.push(expert.up_proj.transpose_for_gemm_gs(gpu, inter, h, routed_gs)?);
+                gate_t.push(
+                    expert
+                        .gate_proj
+                        .transpose_for_gemm_gs(gpu, inter, h, routed_gs)?,
+                );
+                up_t.push(
+                    expert
+                        .up_proj
+                        .transpose_for_gemm_gs(gpu, inter, h, routed_gs)?,
+                );
                 if include_down {
-                    down_t.push(expert.down_proj.transpose_for_gemm_gs(gpu, h, inter, routed_gs)?);
+                    down_t.push(
+                        expert
+                            .down_proj
+                            .transpose_for_gemm_gs(gpu, h, inter, routed_gs)?,
+                    );
                 }
             }
         }
@@ -184,13 +195,12 @@ impl MoeLayer {
 
         // ── Phase A: transpose gate+up routed experts ──
         // ARM-2 Phase-K Family C: native-MXFP4 routed experts are per-32 E8M0.
-        let routed_gs = if self.experts_scale_kind
-            == crate::weight_map::WeightQuantFormat::Mxfp4E8m0
-        {
-            32
-        } else {
-            16
-        };
+        let routed_gs =
+            if self.experts_scale_kind == crate::weight_map::WeightQuantFormat::Mxfp4E8m0 {
+                32
+            } else {
+                16
+            };
         let mut gate_t = Vec::with_capacity(num_experts);
         let mut up_t = Vec::with_capacity(num_experts);
         for expert in &self.weights.experts {
@@ -198,8 +208,16 @@ impl MoeLayer {
                 gate_t.push(QuantizedWeight::null());
                 up_t.push(QuantizedWeight::null());
             } else {
-                gate_t.push(expert.gate_proj.transpose_for_gemm_gs(gpu, inter, h, routed_gs)?);
-                up_t.push(expert.up_proj.transpose_for_gemm_gs(gpu, inter, h, routed_gs)?);
+                gate_t.push(
+                    expert
+                        .gate_proj
+                        .transpose_for_gemm_gs(gpu, inter, h, routed_gs)?,
+                );
+                up_t.push(
+                    expert
+                        .up_proj
+                        .transpose_for_gemm_gs(gpu, inter, h, routed_gs)?,
+                );
             }
         }
         self.gate_ptrs_t = Some(build_ptr_table_from_qw(&gate_t, gpu)?);
@@ -255,7 +273,11 @@ impl MoeLayer {
             if expert.down_proj.is_null() {
                 down_t.push(QuantizedWeight::null());
             } else {
-                down_t.push(expert.down_proj.transpose_for_gemm_gs(gpu, h, inter, routed_gs)?);
+                down_t.push(
+                    expert
+                        .down_proj
+                        .transpose_for_gemm_gs(gpu, h, inter, routed_gs)?,
+                );
             }
         }
         self.down_ptrs_t = Some(build_ptr_table_from_qw(&down_t, gpu)?);
